+++ Game.Api/Models/WebSocket/UnitStateMessageArgs.cs
@@ -6,8 +6,8 @@ namespace Game.Api.Models.WebSocket
     {
         public Guid Id { get; set; }
 
-        public long X { get; set; }
+        public int X { get; set; }
 
-        public long Y { get; set; }
+        public int Y { get; set; }
     }
 }
